Added changes for enable/disable NATS eventing
