+++ src/main/java/io/boomerang/aspect/ActivityEntityUpdateInterceptor.java
@@ -7,13 +7,16 @@ import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.annotation.Before;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Component;
 import io.boomerang.mongo.entity.ActivityEntity;
 import io.boomerang.mongo.repository.FlowWorkflowActivityRepository;
-import io.boomerang.service.EventingService;
+import io.boomerang.service.NATSEventingService;
 
 @Aspect
 @Component
+@ConditionalOnProperty(value = "nats.eventing.enabled", havingValue = "true",
+    matchIfMissing = false)
 public class ActivityEntityUpdateInterceptor {
 
   private static final Logger logger = LogManager.getLogger(ActivityEntityUpdateInterceptor.class);
@@ -22,7 +25,7 @@ public class ActivityEntityUpdateInterceptor {
   FlowWorkflowActivityRepository activityRepository;
 
   @Autowired
-  private EventingService eventingService;
+  private NATSEventingService eventingService;
 
   @Before("execution(* io.boomerang.mongo.repository.FlowWorkflowActivityRepository.save(..))"
       + " && args(entityToBeSaved)")
