+++ src/main/resources/application.properties
@@ -25,6 +25,7 @@ controller.rest.url.base=http://${controller.service.host}
 controller.rest.url.streamlogs=/controller/log/stream
 
 # Eventing properties (NATS Jetstream)
+nats.eventing.enabled=false
 eventing.shared.label.prefix=io.boomerang.event
 eventing.nats.server.urls=nats://bmrg-dev-nats:4222
 eventing.nats.server.reconnect-wait-time=PT10S
